requirements.txt
fastapi
uvicorn
psycopg2-binary
google-generativeai
faker
python-dotenv
pydantic
pytest
requests
httpx
